Added lobby update message server -> client + refactored ClientConnection code to ClientConnectionBase.java
