+++ src/main/java/general/commands/LobbyUpdateBase.java
@@ -1,0 +1,10 @@
+package general.commands;
+
+public class LobbyUpdateBase {
+
+    LobbyUpdateType lobbyUpdateType;
+
+    public LobbyUpdateBase(LobbyUpdateType lobbyUpdateType) {
+        this.lobbyUpdateType = lobbyUpdateType;
+    }
+}
