+++ src/main/java/general/commands/LobbyUpdateType.java
@@ -1,0 +1,9 @@
+package general.commands;
+
+public enum LobbyUpdateType {
+
+    NEW_USER_CONNECTED,
+    DISPLAY_NEXT_QUESTION,
+    SHOW_STANDINGS,
+
+}
