+++ src/main/java/server/PresenterClientConnection.java
@@ -3,15 +3,9 @@ package server;
 import java.io.DataInputStream;
 import java.net.Socket;
 
-public class PresenterClientConnection extends ServerRunnableBase {
-
-    private Socket socket;
-    private String hash;
-    private DataInputStream dataInputStream;
+public class PresenterClientConnection extends ClientConnectionBase {
 
     public PresenterClientConnection(Socket socket, DataInputStream dataInputStream, String hash) {
-        this.socket = socket;
-        this.hash = hash;
-        this.dataInputStream = dataInputStream;
+        super(socket, dataInputStream, hash, "[PlayerClient: " + socket.getInetAddress() + ":" + socket.getPort() + "] ");
     }
 }
