--- src/main/java/server/ServerRunnableBase.java
@@ -1,8 +1,0 @@
-package server;
-
-public class ServerRunnableBase implements Runnable {
-
-    @Override
-    public void run() {
-    }
-}
